+++ app/src/main/java/com/nutricatch/dev/views/factory/PreferencesViewModelFactory.kt
@@ -3,7 +3,7 @@ package com.nutricatch.dev.views.factory
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.nutricatch.dev.data.prefs.Preferences
-import com.nutricatch.dev.views.loading.LoadingViewModel
+import com.nutricatch.dev.views.splash.AppCheckViewModel
 import com.nutricatch.dev.views.navigation.profile.ProfileViewModel
 
 class PreferencesViewModelFactory(private val preferences: Preferences) :
@@ -11,8 +11,8 @@ class PreferencesViewModelFactory(private val preferences: Preferences) :
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         when {
-            modelClass.isAssignableFrom(LoadingViewModel::class.java) -> {
-                return LoadingViewModel(preferences) as T
+            modelClass.isAssignableFrom(AppCheckViewModel::class.java) -> {
+                return AppCheckViewModel(preferences) as T
             }
 
             modelClass.isAssignableFrom(ProfileViewModel::class.java) -> {
